Refactor battle engine to use modular event and turn handling
